refactor(version): Better match cargo-edit layout
