+++ src/ops/version.rs
@@ -2,245 +2,6 @@ use std::str::FromStr;
 
 use crate::error::FatalError;
 
-#[derive(Clone, Debug)]
-pub enum TargetVersion {
-    Relative(BumpLevel),
-    Absolute(semver::Version),
-}
-
-impl TargetVersion {
-    pub fn bump(
-        &self,
-        current: &semver::Version,
-        metadata: Option<&str>,
-    ) -> Result<Option<Version>, FatalError> {
-        let bumped = match self {
-            TargetVersion::Relative(bump_level) => {
-                let mut potential_version = current.to_owned();
-                if bump_level.bump_version(&mut potential_version, metadata)? {
-                    let full_version = potential_version;
-                    let version = Version::from(full_version);
-                    Some(version)
-                } else {
-                    None
-                }
-            }
-            TargetVersion::Absolute(version) => {
-                let mut full_version = version.to_owned();
-                if full_version.build.is_empty() {
-                    if let Some(metadata) = metadata {
-                        full_version.build = semver::BuildMetadata::new(metadata)?;
-                    } else {
-                        full_version.build = current.build.clone();
-                    }
-                }
-                let version = Version::from(full_version);
-                if version.bare_version != Version::from(current.clone()).bare_version {
-                    Some(version)
-                } else {
-                    None
-                }
-            }
-        };
-        Ok(bumped)
-    }
-}
-
-impl Default for TargetVersion {
-    fn default() -> Self {
-        TargetVersion::Relative(BumpLevel::Release)
-    }
-}
-
-impl std::fmt::Display for TargetVersion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
-        match self {
-            TargetVersion::Relative(bump_level) => {
-                write!(f, "{}", bump_level)
-            }
-            TargetVersion::Absolute(version) => {
-                write!(f, "{}", version)
-            }
-        }
-    }
-}
-
-impl std::str::FromStr for TargetVersion {
-    type Err = FatalError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if let Ok(bump_level) = BumpLevel::from_str(s) {
-            Ok(TargetVersion::Relative(bump_level))
-        } else {
-            Ok(TargetVersion::Absolute(semver::Version::parse(s)?))
-        }
-    }
-}
-
-impl clap::builder::ValueParserFactory for TargetVersion {
-    type Parser = TargetVersionParser;
-
-    fn value_parser() -> Self::Parser {
-        TargetVersionParser
-    }
-}
-
-#[derive(Copy, Clone)]
-pub struct TargetVersionParser;
-
-impl clap::builder::TypedValueParser for TargetVersionParser {
-    type Value = TargetVersion;
-
-    fn parse_ref(
-        &self,
-        cmd: &clap::Command,
-        arg: Option<&clap::Arg>,
-        value: &std::ffi::OsStr,
-    ) -> Result<Self::Value, clap::Error> {
-        let inner_parser = TargetVersion::from_str;
-        inner_parser.parse_ref(cmd, arg, value)
-    }
-
-    fn possible_values(
-        &self,
-    ) -> Option<Box<dyn Iterator<Item = clap::builder::PossibleValue> + '_>> {
-        let inner_parser = clap::builder::EnumValueParser::<BumpLevel>::new();
-        #[allow(clippy::needless_collect)] // Erasing a lifetime
-        inner_parser.possible_values().map(|ps| {
-            let ps = ps.collect::<Vec<_>>();
-            let ps: Box<dyn Iterator<Item = clap::builder::PossibleValue> + '_> =
-                Box::new(ps.into_iter());
-            ps
-        })
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Version {
-    pub full_version: semver::Version,
-    pub full_version_string: String,
-    pub bare_version: semver::Version,
-    pub bare_version_string: String,
-}
-
-impl Version {
-    pub fn is_prerelease(&self) -> bool {
-        self.full_version.is_prerelease()
-    }
-}
-
-impl From<semver::Version> for Version {
-    fn from(full_version: semver::Version) -> Self {
-        let full_version_string = full_version.to_string();
-        let mut bare_version = full_version.clone();
-        bare_version.build = semver::BuildMetadata::EMPTY;
-        let bare_version_string = bare_version.to_string();
-        Self {
-            full_version,
-            full_version_string,
-            bare_version,
-            bare_version_string,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, clap::ValueEnum)]
-#[value(rename_all = "kebab-case")]
-pub enum BumpLevel {
-    /// Increase the major version (x.0.0)
-    Major,
-    /// Increase the minor version (x.y.0)
-    Minor,
-    /// Increase the patch version (x.y.z)
-    Patch,
-    /// Increase the rc pre-version (x.y.z-rc.M)
-    Rc,
-    /// Increase the beta pre-version (x.y.z-beta.M)
-    Beta,
-    /// Increase the alpha pre-version (x.y.z-alpha.M)
-    Alpha,
-    /// Remove the pre-version (x.y.z)
-    Release,
-}
-
-impl std::fmt::Display for BumpLevel {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        use clap::ValueEnum;
-
-        self.to_possible_value()
-            .expect("no values are skipped")
-            .get_name()
-            .fmt(f)
-    }
-}
-
-impl std::str::FromStr for BumpLevel {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        use clap::ValueEnum;
-
-        for variant in Self::value_variants() {
-            if variant.to_possible_value().unwrap().matches(s, false) {
-                return Ok(*variant);
-            }
-        }
-        Err(format!("Invalid variant: {}", s))
-    }
-}
-
-impl BumpLevel {
-    pub fn bump_version(
-        self,
-        version: &mut semver::Version,
-        metadata: Option<&str>,
-    ) -> Result<bool, FatalError> {
-        let mut need_commit = false;
-        match self {
-            BumpLevel::Major => {
-                version.increment_major();
-                need_commit = true;
-            }
-            BumpLevel::Minor => {
-                version.increment_minor();
-                need_commit = true;
-            }
-            BumpLevel::Patch => {
-                if !version.is_prerelease() {
-                    version.increment_patch();
-                } else {
-                    version.pre = semver::Prerelease::EMPTY;
-                }
-                need_commit = true;
-            }
-            BumpLevel::Rc => {
-                version.increment_rc()?;
-                need_commit = true;
-            }
-            BumpLevel::Beta => {
-                version.increment_beta()?;
-                need_commit = true;
-            }
-            BumpLevel::Alpha => {
-                version.increment_alpha()?;
-                need_commit = true;
-            }
-            BumpLevel::Release => {
-                if version.is_prerelease() {
-                    version.pre = semver::Prerelease::EMPTY;
-                    need_commit = true;
-                }
-            }
-        };
-
-        if let Some(metadata) = metadata {
-            version.metadata(metadata)?;
-        }
-
-        Ok(need_commit)
-    }
-}
-
 /// Additional version functionality
 pub trait VersionExt {
     /// Increments the major version number for this Version.
